+++ src/render.c
@@ -1,6 +1,6 @@
 #include "./render.h"
 
-STATUS render_render(uint32_t dt) {
+STATUS render_render(float dt) {
   gfx_clear();
   gfx_fill_rect(0, 0, SCREEN_WIDTH, SCREEN_HEIGHT, 0, 0, 0);
   gfx_draw_rect(10, 10, 20, 20, 0, 255, 255);
